backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
pytest==8.1.1
pytest-django==4.8.0
